golang-lint gci fixes
